Wrong method comment
In the comment of the GetInsecureHandler() method it is mentioned to place it in some class of the platform project (android/iOS), but this does not make sense because it causes a circular dependency, since the platform project depends on the shared project. If you are going to use the method created in a class of the platform project, the shared project must contain a reference to the android or iOS project and this is not allowed.